+++ src/trendy_test-reports.ads
@@ -2,6 +2,7 @@ with Trendy_Test;
 
 package Trendy_Test.Reports is
 
+    -- Prints the results of a given test and sets a failure exit code.
     procedure Print_Basic_Report (Results : Trendy_Test.Test_Report_Vectors.Vector);
 
 end Trendy_Test.Reports;